Внесены изменения по комментариям в github.
