+++ ImageFeed/Auth/AuthViewControllerDelegate.swift
@@ -1,0 +1,3 @@
+protocol AuthViewControllerDelegate: AnyObject {
+    func didAuthentificate(_ vc: AuthViewController)
+}
