+++ ImageFeed/Constants.swift
@@ -7,10 +7,3 @@ enum Constants {
     static let accessScope = "public+read_user+write_likes"
     static let defaultBaseURL = URL(string: "https://api.unsplash.com")
 }
-
-class SnakeCaseJSONDecoder: JSONDecoder {
-    override init() {
-        super.init()
-        keyDecodingStrategy = .convertFromSnakeCase
-    }
-}
