+++ ImageFeed/Models/OAuthTokenResponseBody.swift
@@ -1,0 +1,3 @@
+struct OAuthTokenResponseBody: Decodable {
+    let accessToken: String
+}
