+++ ImageFeed/Models/ProfileImage.swift
@@ -1,0 +1,5 @@
+struct ProfileImage: Codable {
+    let small: String
+    let medium: String
+    let large: String
+}
