+++ ImageFeed/Models/ProfileResult.swift
@@ -1,0 +1,6 @@
+struct ProfileResult: Codable {
+    let username: String
+    let firstName: String
+    let lastName: String
+    let bio: String?
+}
