+++ ImageFeed/Models/UserResult.swift
@@ -1,0 +1,3 @@
+struct UserResult: Codable {
+    let profileImage: ProfileImage
+}
