+++ ImageFeed/SnaceCaseDecoder.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+final class SnakeCaseJSONDecoder: JSONDecoder {
+    override init() {
+        super.init()
+        keyDecodingStrategy = .convertFromSnakeCase
+    }
+}
